+++ clang/test/SemaCXX/Inputs/warn-zero-nullptr.h
@@ -1,0 +1,3 @@
+#define NULL (0)
+#define SYSTEM_MACRO (0)
+#define OTHER_SYSTEM_MACRO (NULL)